scripts/requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-genai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
python-pptx>=0.6.23
mammoth>=1.6.0
